Absztraktosítás

+++ 02Nap/02SikidomokTerulete/IPlane.cs
@@ -2,6 +2,6 @@
 {
     public interface IPlane
     {
-        int Area();
+        double Area();
     }
 }
